Revert "Making HIstory Layout"
