+++ lib/constants/color_constants.dart
@@ -1,11 +1,8 @@
 import 'package:flutter/material.dart';
 
 const MainColor = Color.fromRGBO(0, 0, 0, 10);
-const WhiteColor = Color.fromRGBO(255, 255, 255, 10);
-const blueColor = Color.fromRGBO(46, 77, 167, 10);
 const TextColor1 = Color.fromRGBO(63, 61, 86, 10);
 const TextColor2 = Color.fromRGBO(46, 77, 167, 10);
 const textColor2 = Color.fromRGBO(255, 255, 255, 10);
 const textColor3 = Color.fromRGBO(63, 61, 86, 10);
-const textColor4 = Color.fromRGBO(234, 19, 19, 10);
 const BackgroundColor = Color(0xFFFFFF);
