--- lib/screens/history/HistoryScreen.dart
@@ -1,17 +1,0 @@
-import 'package:flutter/material.dart';
-
-class HistoryScreen extends StatefulWidget {
-  HistoryScreen({Key? key}) : super(key: key);
-
-  @override
-  _HistoryScreenState createState() => _HistoryScreenState();
-}
-
-class _HistoryScreenState extends State<HistoryScreen> {
-  @override
-  Widget build(BuildContext context) {
-    return Container(
-      child: Center(child: Text("History Screen"),),
-    );
-  }
-}
