+++ lib/screens/login/components/LoginButton.dart
@@ -1,6 +1,5 @@
 import 'package:bukutamu_android/constants/style_constants.dart';
-import 'package:bukutamu_android/screens/home/HomeScreen.dart';
-import 'package:bukutamu_android/screens/main_screen.dart';
+import 'package:bukutamu_android/screens/mainscreen/MainScreen.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 
@@ -18,7 +17,7 @@ class LoginButton extends StatelessWidget {
             onPressed: () {
               Navigator.push(
                 context,
-                MaterialPageRoute(builder: (context) => main_screen()
+                MaterialPageRoute(builder: (context) => MainScreen()
                 ),
               );
             },
