--- lib/screens/profile/ProfileScreen.dart
@@ -1,19 +1,0 @@
-import 'package:flutter/material.dart';
-
-class ProfileScreen extends StatefulWidget {
-  ProfileScreen({Key? key}) : super(key: key);
-
-  @override
-  _ProfileScreenState createState() => _ProfileScreenState();
-}
-
-class _ProfileScreenState extends State<ProfileScreen> {
-  @override
-  Widget build(BuildContext context) {
-    return Container(
-      child: Center(child: 
-      Text("Profile Screen"),
-      ),
-    );
-  }
-}
