+++ pubspec.yaml
@@ -33,8 +33,6 @@ dependencies:
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.2
-  salomon_bottom_bar: ^3.1.0
-  bottom_navy_bar: ^6.0.0
 
 dev_dependencies:
   flutter_test:
